requirements.txt
traitlets
nodejs
numpy>=1.20,<2
scipy
matplotlib
py_pol
screeninfo
Pillow
numexpr
pandas
ipywidgets
ipympl>=0.9.3
ezdxf 
pyvista
# opencv-python
